Fix hopsize setup
The hopsize was getting setup wrongly. The hopbuffer class expects the
buffersize to be changed to the hopsize. So this gave complications as
that was not implemented and kept receiving the full buffersize. It
worked fine when hopsize and buffersize were equal. That is how it slipt
through.

Now I have tested all methods with buffersizes ranging from 256 to 8192
and hopfactors of 1, 2 and 4. There are more hopfactors available (8 and
16), but I want to only try these on the ones that work well, so I don't
have to test 72 more settings. #timeconsuming

To do:
Find the best looking algorithms, buffersizes and hopfactors and try
them out in a plugin and hear which one works best.