Add hopfactor 8 and 16 for Yin and Yinfast
Yin and Yinfast are the best aubio pitch methods for guitar. So to find
out if a higher hopfactor than 4 is usefull, I made a few more plots of
it using the higher hopfactors.

Hopfactor 16 mostly gave diminishing returns. So depening on the
computer power, it might be wise to go for hopfactor 4 or 8 with
buffersize 2048 or 4096.

With these high hopfactors we got the low latency and low notes
detection, but it still isn't perfect. It probably will never be. It is
quite good though.

To do: implement in Real-Time audio app (eg. DPF) and check if the
latency is really acceptable.