Fix hopbuffer - stable buffersize
When using aubio_module in a DPF audio app, I noticed I couldn't change
the hopfactor. Hopbuffer.cpp expects the buffersize to drop when
increasing the hopfactor. This would mean that a messeage to jack should
be send to lower the bufferingsize. This was to difficult to accomplish
quickly.

So I changed the hopbuffer to keep the buffersize stable when changing
the hopfactor. In fact this means that the buffersize that Jack is using
becomes the hopsize in our app. This might be confusing and will not be
a long term solution. But if you know this, this is easy enough to
understand. That is why we use this for now. Our plan is not to give the
user control over the buffersize and hopfactor anyway (for the
PedalMangler project).

Note: when giving the user control over hopfactor and buffersize there
need to be better name giving or messages to Jack for lowering the
buffersize when increasing the hopfactor.